move comprehension class to comprehensions module
